[mod/#7] LocalHomeImage 로컬 데이터 로드 구조 변경
